+++ controller/src/main/java/org/jboss/as/controller/operations/global/GlobalOperationHandlers.java
@@ -375,7 +375,13 @@ public class GlobalOperationHandlers {
                 context.getResult().set(result);
                 context.completeStep();
             } else {
-                attributeAccess.getReadHandler().execute(context, operation);
+                OperationStepHandler handler = attributeAccess.getReadHandler();
+                ClassLoader oldTccl = SecurityActions.setThreadContextClassLoader(handler.getClass());
+                try {
+                    handler.execute(context, operation);
+                } finally {
+                    SecurityActions.setThreadContextClassLoader(oldTccl);
+                }
             }
         }
     };
@@ -392,7 +398,13 @@ public class GlobalOperationHandlers {
             } else if (attributeAccess.getAccessType() != AccessType.READ_WRITE) {
                 throw new OperationFailedException(new ModelNode().set("Attribute " + attributeName + " is not writeable")); // TODO i18n
             } else {
-                attributeAccess.getWriteHandler().execute(context, operation);
+                OperationStepHandler handler = attributeAccess.getWriteHandler();
+                ClassLoader oldTccl = SecurityActions.setThreadContextClassLoader(handler.getClass());
+                try {
+                    handler.execute(context, operation);
+                } finally {
+                    SecurityActions.setThreadContextClassLoader(oldTccl);
+                }
             }
         }
     };
